crawl-ref/source/webserver/requirements/dev.py2.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    requirements.in/sync.sh
#
appdirs==1.4.4
    # via virtualenv
atomicwrites==1.4.0
    # via pytest
attrs==20.3.0
    # via pytest
backports-abc==0.5
    # via tornado
backports.functools-lru-cache==1.6.1
    # via
    #   isort
    #   wcwidth
click==7.1.2
    # via pip-tools
configparser==4.0.2
    # via
    #   flake8
    #   importlib-metadata
contextlib2==0.6.0.post1
    # via
    #   importlib-metadata
    #   importlib-resources
    #   zipp
coverage==5.4
    # via -r requirements.in/dev.txt
distlib==0.3.1
    # via virtualenv
enum34==1.1.10
    # via flake8
filelock==3.0.12
    # via
    #   tox
    #   virtualenv
flake8-builtins==1.5.3
    # via -r requirements.in/dev.txt
flake8-commas==2.0.0
    # via -r requirements.in/dev.txt
flake8-isort==4.0.0
    # via -r requirements.in/dev.txt
flake8-mutable==1.2.0
    # via -r requirements.in/dev.txt
flake8-polyfill==1.0.2
    # via -r requirements.in/dev.txt
flake8==3.8.4
    # via
    #   -r requirements.in/dev.txt
    #   flake8-builtins
    #   flake8-commas
    #   flake8-isort
    #   flake8-mutable
    #   flake8-polyfill
funcsigs==1.0.2
    # via
    #   mock
    #   pytest
functools32==3.2.3.post2
    # via flake8
futures==3.3.0
    # via
    #   isort
    #   tornado
importlib-metadata==2.1.1
    # via
    #   flake8
    #   pluggy
    #   pytest
    #   tox
    #   virtualenv
importlib-resources==3.3.1
    # via virtualenv
isort==4.3.21
    # via
    #   -r requirements.in/dev.txt
    #   flake8-isort
mccabe==0.6.1
    # via flake8
mock==3.0.5 ; python_version < "3"
    # via -r requirements.in/dev.txt
more-itertools==5.0.0
    # via pytest
packaging==20.9
    # via
    #   pytest
    #   tox
pathlib2==2.3.5
    # via
    #   importlib-metadata
    #   importlib-resources
    #   pytest
    #   virtualenv
pip-tools==5.5.0
    # via -r requirements.in/dev.txt
pluggy==0.13.1
    # via
    #   pytest
    #   tox
py==1.10.0
    # via
    #   pytest
    #   tox
pycodestyle==2.6.0
    # via flake8
pyflakes==2.2.0
    # via flake8
pyparsing==2.4.7
    # via packaging
pytest==4.6.11
    # via -r requirements.in/dev.txt
pyyaml==5.4.1
    # via -r requirements.in/base.txt
scandir==1.10.0
    # via pathlib2
singledispatch==3.4.0.3
    # via
    #   importlib-resources
    #   tornado
six==1.15.0
    # via
    #   mock
    #   more-itertools
    #   pathlib2
    #   pytest
    #   singledispatch
    #   tox
    #   virtualenv
testfixtures==6.17.1
    # via flake8-isort
toml==0.10.2
    # via tox
tornado==5.1.1
    # via -r requirements.in/base.txt
tox==3.21.4
    # via -r requirements.in/dev.txt
typing==3.7.4.3
    # via
    #   flake8
    #   importlib-resources
virtualenv==20.4.2
    # via tox
wcwidth==0.2.5
    # via pytest
zipp==1.2.0
    # via
    #   importlib-metadata
    #   importlib-resources

# The following packages are considered to be unsafe in a requirements file:
# pip
